bugfixes prototypes, format printf, mpq_canonicalize
git-svn-id: svn://scm.gforge.inria.fr/svnroot/apron/trunk@89 d6d8869d-df28-4c9d-b3d8-ddc8926f8a8a

+++ apron/ap_coeff.c
@@ -296,7 +296,7 @@ void ap_coeff_neg(ap_coeff_t* a, const ap_coeff_t* b)
 
 /* Hash */
 
-int ap_coeff_hash(const ap_coeff_t* coeff)
+long ap_coeff_hash(const ap_coeff_t* coeff)
 {
   switch (coeff->discr){
   case AP_COEFF_SCALAR:
